requirements.txt
langchain==0.3.25
langchain-openai==0.3.18
python-dotenv==1.1.0
langchain-community==0.3.24
bs4==0.0.2
unstructured==0.17.2
pypdf==5.5.0
wikipedia==1.4.0
langchain-chroma==0.2.4
pydantic==2.11.5
lxml==5.4.0
tiktoken==0.9.0
pytesseract==0.3.13
openai==1.82.0

# Windows prerequisites: Install Tesseract and Poppler, add to PATH.
